Add Created To GetById Blog Prepare Admin APIS
Add Comment Content Limit Validation

+++ NutrishaAPI/Controllers/V1/Mobile/Comments/CommentController.cs
@@ -1,15 +1,14 @@
 ﻿using System.Threading.Tasks;
-using DL.CommonModels.Paging;
 using DL.DtosV1.Comments;
-using DL.DtosV1.Reactions;
 using DL.ResultModels;
 using DL.Services.Comments;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using NutrishaAPI.Controllers.V1.Bases;
+using NutrishaAPI.Controllers.V1.Mobile.Bases;
+using NutrishaAPI.Validations.Comments;
 using NutrishaAPI.Validations.Shared;
 
-namespace NutrishaAPI.Controllers.V1.Comments
+namespace NutrishaAPI.Controllers.V1.Mobile.Comments
 {
     [Authorize]
     public class CommentController : BaseMobileController
@@ -24,13 +23,14 @@ namespace NutrishaAPI.Controllers.V1.Comments
         [HttpPost("Post")]
         public async Task<IActionResult> PostCommentAsync(PostCommentDto postCommentDto)
         {
-            if (!postCommentDto.IsValidEntityId())
+            var validationResult = postCommentDto.IsValid();
+            if (!validationResult.Success)
             {
-                return InvalidResult(ErrorMessages.InvalidParameters);
+                return InvalidResult(validationResult.Errors);
             }
             
             var result = await _commentService.PostCommentAsync(postCommentDto);
-            return result.Success ? ObjectResult(result.Data) : InvalidResult(result.Errors);
+            return result.Success ? ItemResult(result.Data) : InvalidResult(result.Errors);
         }
 
 
@@ -45,6 +45,8 @@ namespace NutrishaAPI.Controllers.V1.Comments
         [HttpDelete("Delete")]
         public async Task<IActionResult> DeleteCommentAsync(long id)
         {
+            if (id < 1) return InvalidResult(ErrorMessages.InvalidId);
+            
             var result = await _commentService.DeleteCommentAsync(id);
             return result.Success ? EmptyResult() : InvalidResult(result.Errors);
         }
